requirements.txt
openai>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
feedparser>=6.0.10
openai-whisper>=20231117
pydub>=0.25.1

# Video transcript extraction
youtube-transcript-api>=0.6.0

# Web interface dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Semantic search & embeddings
chromadb>=0.4.0
tenacity>=8.2.0
tiktoken>=0.5.0
numpy>=1.24.0

# Background jobs
celery>=5.3.0
redis>=5.0.0

# MCP orchestration layer
mcp>=1.0.0

# Topic clustering
scikit-learn>=1.3.0
anthropic>=0.18.0
